requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.7.0
    # via -r requirements.in
anyio==3.3.4
    # via starlette
appdirs==1.4.4
    # via pooch
asgiref==3.4.1
    # via uvicorn
audioread==2.1.9
    # via librosa
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via soundfile
charset-normalizer==2.0.12
    # via requests
ci-sdr==0.0.0
    # via espnet
click==8.0.3
    # via
    #   nltk
    #   uvicorn
colorama==0.4.4
    # via
    #   click
    #   tqdm
configargparse==1.5.3
    # via espnet
ctc-segmentation==1.7.1
    # via espnet
cython==0.29.24
    # via
    #   ctc-segmentation
    #   pyopenjtalk
    #   pyworld
decorator==5.1.1
    # via librosa
distance==0.1.3
    # via g2p-en
einops==0.4.0
    # via ci-sdr
espnet==0.10.6
    # via -r requirements.in
espnet-tts-frontend==0.0.3
    # via espnet
fastapi==0.70.0
    # via -r requirements.in
filelock==3.4.2
    # via espnet
g2p-en==2.1.0
    # via espnet-tts-frontend
h11==0.12.0
    # via uvicorn
h5py==3.6.0
    # via espnet
humanfriendly==10.0
    # via espnet
idna==3.3
    # via
    #   anyio
    #   requests
inflect==5.4.0
    # via
    #   espnet-tts-frontend
    #   g2p-en
jaconv==0.3
    # via espnet-tts-frontend
jamo==0.4.1
    # via espnet
joblib==1.1.0
    # via
    #   librosa
    #   nltk
    #   scikit-learn
kaldiio==2.17.2
    # via espnet
librosa==0.9.0
    # via espnet
llvmlite==0.38.0
    # via numba
nltk==3.7
    # via
    #   espnet
    #   g2p-en
numba==0.55.1
    # via
    #   librosa
    #   resampy
numpy==1.20.0
    # via
    #   -r requirements.in
    #   ci-sdr
    #   ctc-segmentation
    #   g2p-en
    #   h5py
    #   kaldiio
    #   librosa
    #   numba
    #   pyopenjtalk
    #   pytorch-wpe
    #   pyworld
    #   resampy
    #   scikit-learn
    #   scipy
    #   torch-complex
packaging==21.3
    # via
    #   librosa
    #   pooch
pooch==1.6.0
    # via librosa
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via fastapi
pyopenjtalk @ git+https://github.com/VOICEVOX/pyopenjtalk@a85521a0a0f298f08d9e9b24987b3c77eb4aaff5
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
pypinyin==0.44.0
    # via
    #   espnet
    #   espnet-tts-frontend
pyreadline3==3.4.1
    # via humanfriendly
python-multipart==0.0.5
    # via -r requirements.in
pytorch-wpe==0.0.1
    # via espnet
pyworld==0.3.0
    # via espnet
pyyaml==6.0
    # via
    #   -r requirements.in
    #   espnet
regex==2022.1.18
    # via nltk
requests==2.27.1
    # via pooch
resampy==0.2.2
    # via librosa
scikit-learn==1.0.2
    # via librosa
scipy==1.7.1
    # via
    #   -r requirements.in
    #   ci-sdr
    #   espnet
    #   librosa
    #   resampy
    #   scikit-learn
sentencepiece==0.1.96
    # via espnet
six==1.16.0
    # via
    #   pyopenjtalk
    #   python-multipart
    #   resampy
sniffio==1.2.0
    # via anyio
soundfile==0.10.3.post1
    # via
    #   -r requirements.in
    #   espnet
    #   librosa
starlette==0.16.0
    # via fastapi
threadpoolctl==3.1.0
    # via scikit-learn
torch==1.10.2
    # via
    #   ci-sdr
    #   espnet
torch-complex==0.2.1
    # via espnet
tqdm==4.62.3
    # via
    #   nltk
    #   pyopenjtalk
typeguard==2.13.3
    # via espnet
typing-extensions==3.10.0.2
    # via
    #   pydantic
    #   torch
unidecode==1.3.2
    # via espnet-tts-frontend
urllib3==1.26.8
    # via requests
uvicorn==0.15.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
